Refactor : 셀프 리팩토링, reset, common scss 적용
