+++ src/index.js
@@ -2,6 +2,8 @@ import React from 'react';
 // import { library, icon } from '@fortawesome/fontawesome-svg-core'
 import ReactDOM from 'react-dom';
 import './index.css';
+import './styles/common.scss';
+import './styles/reset.scss';
 import Routes from './Routes';
 
 ReactDOM.render(
