--- src/pages/namsunkim/Feed/Components/FeedNav.js
@@ -1,10 +1,0 @@
-import React, { Component } from 'react';
-
-export default class FeedNav extends Component {
-  render() {
-    return (
-      <>
-      </>
-    );
-  }
-}
