+++ src/pages/namsunkim/Feed/Feed.scss
@@ -1,4 +1,4 @@
-@import "../styles/common.scss";
+@import '../styles/variable.scss';
 
 .feed {
     width: 100%;
