+++ src/pages/namsunkim/Main/Main.scss
@@ -1,6 +1,6 @@
 @import url('https://fonts.googleapis.com/css2?family=Lobster&display=swap');
 @import url('//cdn.jsdelivr.net/npm/xeicon@2.3.3/xeicon.min.css');
-@import "../styles/common.scss";
+@import '../styles/variable.scss';
 
 a, span {
     font-size: 0.875rem;
