+++ src/pages/namsunkim/Nav/Components/NavProfile/NavProfile.scss
@@ -1,4 +1,4 @@
-@import '../../styles/common.scss';
+@import '../../../styles/variable.scss';
 
 .profile-wrap {
   padding-top: 0.3125rem;
