+++ src/pages/namsunkim/Recommend/Components/RecommendAccount.scss
@@ -1,4 +1,4 @@
-@import '../../styles/common.scss';
+@import '../../styles/variable.scss';
 
 .recommendAccount {
   display: flex;
