+++ src/pages/namsunkim/Recommend/Recommend.scss
@@ -1,5 +1,4 @@
-// @import "../../styles/common.scss";
-@import "../styles/common.scss";
+@import '../styles/variable.scss';
 
 .main-right {
     width: 18.31rem;
