+++ src/pages/namsunkim/Reply/Reply.scss
@@ -1,4 +1,4 @@
-@import "../styles/common.scss";
+@import '../styles/variable.scss';
 
 .reply-list{
     padding: 0 1rem 0.75rem 1rem;
