+++ src/pages/namsunkim/Story/Components/oneStory.scss
@@ -1,0 +1,32 @@
+@import '../../styles/variable.scss';
+@import '../../../../styles/common.scss';
+
+
+li {
+  margin: 0 0.625rem;
+
+  .gradient-wrap {
+    @extend %flexCenter;
+      width: 3.75rem;
+      height: 3.75rem;
+      margin-bottom: 0.3125rem;
+      border-radius: 100%;
+      background: radial-gradient(circle at bottom left, #F58529 20%, #C42D91);
+
+      img {
+          border-radius: 50%;
+          width: 3.5rem;
+          border: 2px solid $white;
+      }
+  }
+
+  p {
+      text-overflow: ellipsis;
+      max-width: 3.75rem;
+      overflow: hidden;
+      margin: 0;
+      font-size: 0.75rem;
+      letter-spacing: 0.1px;
+      text-align: center;
+  }
+}
